refactor: remove react-hook-form dependency and migrate CronInput component
