+++ packages/frontend/src/components/CronInput/DailyInputs.tsx
