+++ packages/frontend/src/components/CronInput/WeeklyInputs.tsx
